+++ lib/wavefront-sdk/core/exception.rb
@@ -28,6 +28,7 @@ module Wavefront
     class InvalidIntegrationId < RuntimeError; end
     class InvalidLinkTemplate < RuntimeError; end
     class InvalidMaintenanceWindowId < RuntimeError; end
+    class InvalidMonitoredClusterId < RuntimeError; end
     class InvalidMessageId < RuntimeError; end
     class InvalidMetricName < RuntimeError; end
     class InvalidMetricValue < RuntimeError; end
@@ -38,6 +39,7 @@ module Wavefront
     class InvalidPrefixLength < RuntimeError; end
     class InvalidProxyId < RuntimeError; end
     class InvalidRelativeTime < RuntimeError; end
+    class InvalidSamplingValue < RuntimeError; end
     class InvalidSavedSearchEntity < RuntimeError; end
     class InvalidSavedSearchId < RuntimeError; end
     class InvalidServiceAccountId < RuntimeError; end
@@ -51,6 +53,7 @@ module Wavefront
     class InvalidUserGroupId < RuntimeError; end
     class InvalidVersion < RuntimeError; end
     class InvalidWebhookId < RuntimeError; end
+    class NetworkTimeout < RuntimeError; end
     class NotImplemented < RuntimeError; end
     class SocketError < RuntimeError; end
     class UnparseableResponse < RuntimeError; end